Refactor models/viewmodels
