--- src/XboxBatteryMonitor/Models/ControllerInfo.cs
@@ -1,12 +1,0 @@
-using CommunityToolkit.Mvvm.ComponentModel;
-
-namespace XboxBatteryMonitor.Models;
-
-public partial class ControllerInfo : ObservableObject
-{
-    [ObservableProperty]
-    private string name = "Xbox Controller";
-
-    [ObservableProperty]
-    private BatteryInfo batteryInfo = new();
-}
